Use the types to ensure the right thing is deleted.

+++ projects/plugins/boost/app/modules/cache/pre-wordpress/storage/File_Storage.php
@@ -113,9 +113,9 @@ class File_Storage implements Storage {
 		error_log( "invalidate: $path $type" ); // phpcs:ignore WordPress.PHP.DevelopmentFunctions.error_log_error_log
 		$normalized_path = $this->root_path . Boost_Cache_Utils::normalize_request_uri( $path );
 
-		if ( is_dir( $normalized_path ) ) {
+		if ( in_array( $type, array( Boost_Cache_Utils::DELETE_FILES, Boost_Cache_Utils::DELETE_ALL ), true ) && is_dir( $normalized_path ) ) {
 			return Boost_Cache_Utils::delete_directory( $normalized_path, $type );
-		} else {
+		} elseif ( $type === Boost_Cache_Utils::DELETE_FILE ) {
 			return Filesystem_Utils::delete_file( $normalized_path );
 		}
 	}
